+++ src/extractdata/entities/IncidentDetail.java
@@ -16,8 +16,8 @@ public class IncidentDetail {
     private String ds_name;
     private int key;
     private int type;
-    private float latitude;
-    private float longitude;
+    private double latitude;
+    private double longitude;
     private float acc_x;
     private float acc_y;
     private float acc_z;
@@ -39,11 +39,11 @@ public class IncidentDetail {
         return type;
     }
 
-    public float getLatitude() {
+    public double getLatitude() {
         return latitude;
     }
 
-    public float getLongitude() {
+    public double getLongitude() {
         return longitude;
     }
 
@@ -91,11 +91,11 @@ public class IncidentDetail {
         this.type = type;
     }
     
-    public void setLatitude(float latitude) {
+    public void setLatitude(double latitude) {
         this.latitude = latitude;
     }
 
-    public void setLongitude(float longitude) {
+    public void setLongitude(double longitude) {
         this.longitude = longitude;
     }
